+++ mirrord/config/src/agent.rs
@@ -84,6 +84,12 @@ pub struct AgentConfig {
         unstable
     )]
     pub flush_connections: bool,
+
+    /// Create an agent that returns an error after accepting the first client. For testing
+    /// purposes. Only supported with job agents (not with ephemeral agents).
+    #[cfg(all(debug_assertions, not(test)))] // not(test) so that it's not included in the schema json.
+    #[config(env = "MIRRORD_AGENT_TEST_ERROR", default = false, unstable)]
+    pub test_error: bool,
 }
 
 #[cfg(test)]